requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.6
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
editor==1.6.6
exceptiongroup==1.2.1
fastembed==0.2.7
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-tools==1.64.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
inquirer==3.2.4
joblib==1.4.2
llama-index==0.10.39
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.39.post1
llama-index-embeddings-fastembed==0.1.4
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.21
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-qdrant==0.1.3
llama-parse==0.4.3
llamaindex-py-client==0.1.19
loguru==0.7.2
markdown-it-py==3.0.0
marshmallow==3.21.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
onnx==1.16.1
onnxruntime==1.18.0
openai==1.30.3
packaging==24.0
pandas==2.0.3
pillow==10.3.0
portalocker==2.8.2
protobuf==5.27.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
readchar==4.1.0
regex==2024.5.15
requests==2.32.2
rich==13.7.1
runs==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
striprtf==0.0.26
sympy==1.12
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.15.2
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
xmod==1.8.1
yarl==1.9.4
